+++ app/controllers/kvstore_admin_controller.rb
@@ -1,6 +1,4 @@
 class KvstoreAdminController < AdminController
-  http_basic_authenticate_with name: Figaro.env.http_basic_username, password: Figaro.env.http_basic_password
-
   def index
     @kvstores = Kvstore.all.page(params[:page])
   end